Lots of interconnected api breaking changes
Making "dependents" dictionaries instead of parallel lists
Moving interpolators to be separate, only 1 in - 1 out, owned by
Dependents
Adding the differential flag to dependents (not used yet, as `__add__` is
not yet implemented)
Make "Setpoints" object which shares a parent with Dependents, for now
called "Variable", though I don't like that name since it overloads a
term used in wt. It is not exported into any other namespace, and is not
intended to be used directly. At this time, Setpoints does not do
anything in addition to what Variable does, but I have archietected it
in a way that it _can_ diverge without affecting Dependents.

Update io for base curve and topas to reflect api changes, add new
attributes to the .curve format.

Tests updated with api changes, namely dependents being dictionaries and
setpoints being objects, not bare arrays.

Uses `__call__` rather than defining a separate function for
interpolators.